\name{lsrm-package}
\alias{lsrm-package}
\alias{lsrm}
\docType{package}
\title{
\packageTitle{lsrm}
}
\description{
\packageDescription{lsrm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lsrm}
\packageIndices{lsrm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lsrm}

Maintainer: \packageMaintainer{lsrm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
