% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsrm-package.R, R/lsrm.R
\docType{package}
\name{lsrm}
\alias{lsrm}
\alias{lsrm-package}
\title{Title}
\usage{
lsrm(data, nsample, nitem, ndim = 3, niter = 30000, nburn = 5000,
  nthin = 5, nprint = 100, jump_beta = 0.1, jump_theta = 0.5,
  jump_z = 0.05, jump_w = 0.05, pr_mean_beta = 0, pr_sd_beta = 10,
  pr_mean_theta = 0, pr_sd_theta = 10, pr_mean_z = 0, pr_sd_z = 1,
  pr_mean_w = 0, pr_sd_w = 1, pr_z_times = 4, option = TRUE,
  dist_option = 1, cores = 1)
}
\description{
Main Description
}
