% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsrm_ms.R
\name{lsrm_ms}
\alias{lsrm_ms}
\title{title}
\usage{
lsrm_ms(dataset, nsample, nitem, ndim = 3, nset, nsamp_max, nitem_max,
  ntotal_max, niter = 30000, nburn = 5000, nthin = 5, nprint = 100,
  jump_beta = 0.2, jump_theta = 1, jump_z = 0.05, jump_w = 0.05,
  pr_mean_beta = 0, pr_sd_beta = 10, pr_mean_theta = 0,
  pr_sd_theta = 10, pr_mean_z = 0, pr_sd_z = 1, pr_mean_w = 0,
  pr_sd_w = 1, pr_z_times = 4, option = TRUE, latent_equal = TRUE,
  cores = 1)
}
\description{
title
}
